+++ lib/src/builder/create/mainbuilder.dart
@@ -1,5 +1,5 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:dart_style/dart_style.dart';
+import 'package:dart_style/dart_style.dart' show DartFormatter;
 
 /// Simple Function to render the `main` pheasant file entry.
 ///