+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0-beta
+- Added Support for Event Handling
+- Added basic state management support
+- Added support for passing data to child.
+- Extended support for custom components via slots
+
 ## 0.1.0-alpha
 
 - Initial version.